+++ lib/games/dota2.js
