+++ lib/games/leagueoflegends/championgg.js
