+++ lib/games/leagueoflegends/help.js
