+++ lib/games/leagueoflegends/index.js
