+++ lib/games/leagueoflegends/riot.js
